Property Detail Page 1. Search and Filter Completed
Ongoing
1. Owner Inquiry List, Delete Page, Display Review
2. User Task
